fix_bug_delete_materia

+++ BackEnd/src/materia/materia.controller.ts
@@ -20,56 +20,82 @@ function inputS (req: Request, res: Response, next: NextFunction) {
 }
 
 async function findAll(req: Request, res: Response) {
+    const em = orm.em.fork();
     try{
         const materias = await em.find(Materia, {})
         res.header('Access-Control-Allow-Origin', '*');
         res.status(200).json(materias);
     } catch (error:any){
-    res.status(500).json({ mensaje: error.message });
+        console.error('Error al obtener materia:', error);
+        res.status(500).json({ mensaje: error.message });
     }
 }
 
 async function findOne (req:Request, res:Response) {
+    const em = orm.em.fork();
+    const id = Number.parseInt(req.params.id)
     try{
-        const id = Number.parseInt(req.params.id)
-        const materia = await em.findOneOrFail(Materia,{ id })
+        const materia = await em.findOne(Materia, { id })
+        if(!materia){
+            return res.status(404).json({ mensaje: 'Materia no encontrado' })
+        }
         res.header('Access-Control-Allow-Origin', '*');
         res.status(200).json(materia);
     } catch (error:any){
+        console.error('Error al buscar el materia:', error);
         res.status(500).json({ mensaje: error.message })
     }
 }
 
 async function add (req:Request, res:Response) { 
+    const em = orm.em.fork();
+    const input = req.body.inputS
+
     try{
-        const materia = em.create(Materia, req.body.inputS)
-        await em.flush()
+        const nuevaMateria = em.create(Materia, input)
+        await em.persistAndFlush(nuevaMateria)
         res.header('Access-Control-Allow-Origin', '*');
-        res.status(201).json({ mensaje: 'Materia creado', data: materia});
+        res.status(201).json({ mensaje: 'Materia creado', data: nuevaMateria});
     } catch (error:any){
+        console.error('Error al agregar materia:', error);
         res.status(500).json({ mensaje: error.message });
     }
 }
 
 async function update(req:Request, res:Response) {
+    const em = orm.em.fork();
+    const id = parseInt(req.params.id)
+    const input = req.body.inputS
+
     try{
-        const id = Number.parseInt(req.params.id)
-        const materiaToUpdate = await em.findOneOrFail(Materia, { id })
-        em.assign(materiaToUpdate, req.body.inputS)
+        const materia = await em.findOne(Materia, { id })
+        if(!materia){
+            return res.status(404).json({ mensaje: 'Materia no encontrado' })
+        }
+        em.assign(materia, input);
         await em.flush()
         res.header('Access-Control-Allow-Origin', '*');
-        res.status(200).json({ mensaje: 'Materia actualizado', data: materiaToUpdate});
+        res.status(200).json({ mensaje: 'Materia actualizado', data: materia});
     } catch (error:any){
+        console.error('Error al actualizar materia:', error);
         res.status(500).json({ mensaje: error.message });
     }
 }
 
 async function remove(req:Request, res:Response){
-    try{
-        const id = Number.parseInt(req.params.id)
-        const materia = em.getReference(Materia, id)
+    const em = orm.em.fork();
+    const id = Number.parseInt(req.params.id)
+    try{ 
+        const materia = await em.findOne(Materia, { id })
+        if(!materia){
+            return res.status(404).json({ mensaje: 'Materia no encontrado' })
+        }
+
         await em.removeAndFlush(materia)
+        res.header('Access-Control-Allow-Origin', '*');
+        res.status(200).json({ mensaje: 'Materia eliminado'});
     } catch (error:any){
+        console.error('Error al eliminar materia:', error);
         res.status(500).json({ mensaje: error.message });
     }
 }
